+++ packages/client/src/lib/simulation/updates/EpisodeUpdate.ts
@@ -1,0 +1,22 @@
+import { d } from "@/lib/drizzle";
+
+/**
+ * An episode (i.e. predefined) update.
+ */
+export class EpisodeUpdate {
+  static is(obj: any): obj is EpisodeUpdate {
+    return obj instanceof EpisodeUpdate;
+  }
+
+  constructor(
+    readonly id: string,
+    readonly parentId: string | null,
+    readonly episodeId: string,
+    readonly chunkIndex: number,
+    readonly text: string,
+    readonly directorUpdate: Pick<
+      typeof d.directorUpdates.$inferSelect,
+      "id" | "code" | "createdAt"
+    > | null,
+  ) {}
+}